refactor: poc server actions and cache
