+++ app/(connected)/(with-sidebar)/(with-chat)/mes-jeunes/[idJeune]/actions/[idAction]/DetailActionPage.tsx
@@ -64,7 +64,9 @@ function DetailActionPage({
   const suppressionModalRef = useRef<ModalHandles>(null)
 
   async function updateStatutAction(statutChoisi: StatutAction): Promise<void> {
-    const { modifierAction } = await import('services/actions.service')
+    const { modifierAction } = await import(
+      'server-actions/actions.server-actions'
+    )
     await modifierAction(action.id, { statut: statutChoisi })
     setStatut(statutChoisi)
 
@@ -179,7 +181,7 @@ function DetailActionPage({
           </h2>
           {!qualifiee && (
             <ButtonLink
-              href={`/mes-jeunes/${jeune.id}/actions/${action.id}/modification?misc=${random}`}
+              href={`/mes-jeunes/${jeune.id}/actions/${action.id}/modification`}
               style={ButtonStyle.SECONDARY}
             >
               <IconComponent
